@@ -1,6 +1,6 @@
 {
   "name": "jolc",
-  "version": "2.0.3",
+  "version": "2.0.4",
   "scripts": {
     "start": "python3 -m uvicorn api.main:app --host=0.0.0.0 --port=${PORT:-5000}",
     "dev:client": "cd client && vite",
@@ -9,24 +9,25 @@
     "serve": "cd client && vite preview"
   },
   "dependencies": {
-    "@svgr/plugin-jsx": "^8.1.0",
-    "@svgr/plugin-svgo": "^8.1.0",
-    "ace-builds": "1.13.0",
-    "axios": "1.1.3",
+    "@svgr/plugin-jsx": "8.1.0",
+    "@svgr/plugin-svgo": "8.1.0",
+    "@vitejs/plugin-react": "4.2.1",
+    "ace-builds": "1.32.3",
+    "axios": "1.6.4",
     "graphviz-react": "1.2.5",
-    "react": "17.0.0",
+    "react": "18.2.0",
     "react-ace": "10.1.0",
-    "react-dom": "17.0.0",
-    "react-redux": "8.0.5",
+    "react-dom": "18.2.0",
+    "react-redux": "9.0.4",
     "react-split-grid": "1.0.4",
-    "redux-devtools-extension": "2.13.9",
-    "wouter": "2.8.1"
+    "redux": "5.0.1",
+    "wouter": "2.12.1"
   },
   "devDependencies": {
-    "@vitejs/plugin-react-refresh": "1.3.6",
-    "eslint-plugin-react": "7.31.10",
-    "vite": "3.2.3",
-    "vite-plugin-svgr": "^4.2.0"
+    "eslint-plugin-react": "7.33.2",
+    "vite": "5.0.10",
+    "vite-plugin-svgr": "4.2.0"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "type": "module"
 }
