@@ -1,15 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
 import '@/styles/index.css'
-import App from '@/components/App'
+import React from 'react'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
+
+import App from '@/components/App'
 import store from '@/store'
 
-ReactDOM.render(
+const mainElement = document.querySelector('main')
+const root = createRoot(mainElement)
+
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
